requirements.txt
google-generativeai
flask
requests
